diarization/requirements.txt
torchmetrics==1.9.0
huggingface_hub==0.36.2
pyannote.audio==4.0.4
lightning==2.6.1
torchvision==0.25.0
fastapi==0.135.1
uvicorn==0.42.0
loguru==0.7.3
numpy==2.4.3
soundfile
python-multipart==0.0.22
